sizehelper, blockchain: move test code for CalcNumEntries
The test code for CalcNumEntries is moved and is made better by adding
more test cases for it. This catches the bug where the current code
isn't accurate to 1% of the requested maxSize.
